Align m_id and o_id handling with NetHack 3.7
Trying something to fix the "cant find o_id <number>" error. There is no
reason the m_id/o_id handling should be different from vanilla.

This fixes a potential issue in nextoid(), sets the type back, and
removes unnecessary if blocks.

+++ include/monst.h
@@ -71,7 +71,7 @@ enum m_ap_types {
 struct monst {
     struct monst *nmon;
     struct permonst *data;
-    size_t m_id;
+    unsigned m_id;
     short mnum;           /* permanent monster index number */
     short cham;           /* if shapeshifter, orig mons[] idx goes here */
     short movement;       /* movement points (derived from permonst definition and added effects */
